backend/requirements.txt
fastapi
uvicorn
websockets
langgraph
langchain-groq
langchain-community
tavily-python
python-dotenv
pydantic
asyncpg
langchain-huggingface
sentence-transformers
langgraph-checkpoint-postgres
psycopg[binary,pool]